// world-cup-api/src/match/model/match-event-type.enum.ts
export enum MatchEventType {
  KICKOFF_EVENT = 'KICKOFF_EVENT',
  HALF_TIME_EVENT = 'HALF_TIME_EVENT',
  BALL_POSSESSION_EVENT = 'BALL_POSSESSION_EVENT',
  ATTACK_EVENT = 'ATTACK_EVENT',
  DEFENSE_EVENT = 'DEFENSE_EVENT',
  FREE_KICK_FOR_EVENT = 'FREE_KICK_FOR_EVENT',
  FREE_KICK_AGAINST_EVENT = 'FREE_KICK_AGAINST_EVENT',
  PENALTY_FOR_EVENT = 'PENALTY_FOR_EVENT',
  PENALTY_AGAINST_EVENT = 'PENALTY_AGAINST_EVENT',
  CORNER_FOR_EVENT = 'CORNER_FOR_EVENT',
  CORNER_AGAINST_EVENT = 'CORNER_AGAINST_EVENT',
  THROW_IN_FOR_EVENT = 'THROW_IN_FOR_EVENT',
  THROW_IN_AGAINST_EVENT = 'THROW_IN_AGAINST_EVENT',
  YELLOW_CARD_EVENT = 'YELLOW_CARD_EVENT',
  RED_CARD_EVENT = 'RED_CARD_EVENT',
  LAST_PLAY_ONE_ON_ONE_FOR_EVENT = 'LAST_PLAY_ONE_ON_ONE_FOR_EVENT',
  LAST_PLAY_ONE_ON_ONE_AGAINST_EVENT = 'LAST_PLAY_ONE_ON_ONE_AGAINST_EVENT',
  LAST_PLAY_PENALTY_FOR_EVENT = 'LAST_PLAY_PENALTY_FOR_EVENT',
  LAST_PLAY_PENALTY_AGAINST_EVENT = 'LAST_PLAY_PENALTY_AGAINST_EVENT',
  FORFEIT_EVENT = 'FORFEIT_EVENT',
  END = 'END',
}